[ mail ] new script to show new mails in inbox
this script assumes there is a global maildir, e.g. `~/.Mail` and
searches in every folder in `~/.Mail` for a folder named `Inbox` and in
each inbox-folder for a folder named `new`, e.g.:

`~/.Mail/private/Inbox/new` or
`~/.Mail/work/Inbox/new` or
`~/.Mail/com.google.bettydavis/Inbox/new`

in each of these `new` folders it counts the mails and the sum of these
will be echoed.